x/wms/types/genesis_test.go
package types_test

import (
	"testing"

	"kewr/x/wms/types"

	"github.com/stretchr/testify/require"
)

func TestGenesisState_Validate(t *testing.T) {
	tests := []struct {
		desc     string
		genState *types.GenesisState
		valid    bool
	}{
		{
			desc:     "default is valid",
			genState: types.DefaultGenesis(),
			valid:    true,
		},
		{
			desc: "valid genesis state",
			genState: &types.GenesisState{

				WarehouseList: []types.Warehouse{
					{
						Id: 0,
					},
					{
						Id: 1,
					},
				},
				WarehouseCount: 2,
				// this line is used by starport scaffolding # types/genesis/validField
			},
			valid: true,
		},
		{
			desc: "duplicated warehouse",
			genState: &types.GenesisState{
				WarehouseList: []types.Warehouse{
					{
						Id: 0,
					},
					{
						Id: 0,
					},
				},
			},
			valid: false,
		},
		{
			desc: "invalid warehouse count",
			genState: &types.GenesisState{
				WarehouseList: []types.Warehouse{
					{
						Id: 1,
					},
				},
				WarehouseCount: 0,
			},
			valid: false,
		},
		// this line is used by starport scaffolding # types/genesis/testcase
	}
	for _, tc := range tests {
		t.Run(tc.desc, func(t *testing.T) {
			err := tc.genState.Validate()
			if tc.valid {
				require.NoError(t, err)
			} else {
				require.Error(t, err)
			}
		})
	}
}
